+++ lib/src/controls/navigation/navigation_view/body.dart
@@ -185,6 +185,7 @@ class NavigationViewContext extends InheritedWidget {
     required this.previousItemIndex,
     required this.isTransitioning,
     required this.toggleButtonPosition,
+    required this.canPop,
     super.key,
   });
 
@@ -200,6 +201,7 @@ class NavigationViewContext extends InheritedWidget {
       previousItemIndex: parent.previousItemIndex,
       isTransitioning: parent.isTransitioning,
       toggleButtonPosition: parent.toggleButtonPosition,
+      canPop: parent.canPop,
       child: child,
     );
   }
@@ -231,6 +233,9 @@ class NavigationViewContext extends InheritedWidget {
   /// The position of the toggle pane button.
   final PaneToggleButtonPosition toggleButtonPosition;
 
+  /// Whether the navigation view can pop the current item.
+  final bool canPop;
+
   /// Returns the closest [NavigationViewContext] ancestor, if any.
   static NavigationViewContext? maybeOf(BuildContext context) {
     return context.dependOnInheritedWidgetOfExactType<NavigationViewContext>();
@@ -251,6 +256,7 @@ class NavigationViewContext extends InheritedWidget {
         oldWidget.pane != pane ||
         oldWidget.previousItemIndex != previousItemIndex ||
         oldWidget.isTransitioning != isTransitioning ||
-        oldWidget.toggleButtonPosition != toggleButtonPosition;
+        oldWidget.toggleButtonPosition != toggleButtonPosition ||
+        oldWidget.canPop != canPop;
   }
 }